Due to a miscount in comp_options_by_version, comp_soul and comp_maskedanim weren't actually ever optional! They weren't being counted as options in any complevel, so their values were always forced.
This was fixed by rewriting G_Compatibility to specify the complevel at which
each change becomes optional. This is in some sense the reverse of the previous
code; instead of a number of options per level, we specify a level for each
option. I believe, this way round, it will be easier to maintain.