Embarrassing bug fix: proper use of AC_ARG_WITH
It turns out I've been doing this all wrong for years. The syntax is
  AC_ARG_WITH(library, help, action-if-given, action-if-not-given)
It is _not_:
  AC_ARG_WITH(library, help, action-if-enabled, action-if-disabled)

This is the reason for the "bug" where --with-blah and --without-blah
do the same thing, but a different thing from both being omitted.

What should be done: (and similarly for AC_ARG_ENABLE)
- leave action-if-given empty. AC_ARG_WITH thus sets with_library
  to "yes" (--with-library used) or "no" (--without-library used.)
- supply a default with_library "yes" or "no" in action-if-not-given.
- test the value of with_library, and perform actions accordingly,
  as a separate step, not embedded within the body of AC_ARG_WITH.
  Can also use m4sh AS_IF instead of plain sh conditionals here.

http://www.flameeyes.eu/autotools-mythbuster/autoconf/arguments.html